chatsystem/backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9   # used by Alembic offline mode

# Redis
redis[asyncio]>=5.0.0

# AI / LangChain
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0

# Auth
bcrypt>=3.2.0,<4.0
python-jose[cryptography]>=3.3.0

# Settings
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0

# HTTP client (WhatsApp Cloud API)
httpx>=0.27.0

# WebSockets (bundled with uvicorn[standard], listed for clarity)
websockets>=12.0
